Permissions moved to controller in user

+++ src/main/java/pl/skyterix/sadsky/user/CommandUserController.java
@@ -13,8 +13,13 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
+import pl.skyterix.sadsky.exception.Errors;
+import pl.skyterix.sadsky.exception.GroupUnauthorizedException;
+import pl.skyterix.sadsky.user.domain.User;
 import pl.skyterix.sadsky.user.domain.UserFacade;
 import pl.skyterix.sadsky.user.domain.dto.UserDTO;
+import pl.skyterix.sadsky.user.domain.group.Permission;
+import pl.skyterix.sadsky.user.domain.group.SelfPermission;
 import pl.skyterix.sadsky.user.domain.group.strategy.GroupStrategy;
 import pl.skyterix.sadsky.user.request.UserDetailsRequest;
 import pl.skyterix.sadsky.user.request.UserReplaceRequest;
@@ -46,28 +51,52 @@ class CommandUserController implements CommandUserControllerPort {
     @Override
     @DeleteMapping("/{userId}")
     public void deleteUser(@PathVariable UUID userId) {
-        userFacade.deleteUser(userId);
+        User currentUser = userFacade.getAuthenticatedUser();
+
+        // Checks if currentUser has permission to delete that user
+        if (currentUser.hasPermission(userId, SelfPermission.DELETE_SELF_USER, Permission.DELETE_USER)) {
+            userFacade.deleteUser(userId);
+        } else
+            throw new GroupUnauthorizedException(Errors.UNAUTHORIZED_GROUP.getErrorMessage(currentUser.getGroup().getName()));
     }
 
     @Override
     @PatchMapping("/{userId}")
     public void updateUser(@PathVariable UUID userId, @RequestBody @Validated UserUpdateRequest userUpdateRequest) {
+        User currentUser = userFacade.getAuthenticatedUser();
+
         UserDTO userDTO = jpaModelMapper.mapEntity(userUpdateRequest, UserDTO.class);
 
-        userFacade.updateUser(userId, userDTO);
+        // Checks if currentUser has permission to update that user
+        if (currentUser.hasPermission(userId, SelfPermission.UPDATE_SELF_USER, Permission.UPDATE_USER)) {
+            userFacade.updateUser(userId, userDTO);
+        } else
+            throw new GroupUnauthorizedException(Errors.UNAUTHORIZED_GROUP.getErrorMessage(currentUser.getGroup().getName()));
     }
 
     @Override
     @PutMapping("/{userId}")
     public void replaceUser(@PathVariable UUID userId, @RequestBody @Validated UserReplaceRequest userReplaceRequest) {
+        User currentUser = userFacade.getAuthenticatedUser();
+
         UserDTO userDTO = jpaModelMapper.mapEntity(userReplaceRequest, UserDTO.class);
 
-        userFacade.replaceUser(userId, userDTO);
+        // Checks if currentUser has permission to replace that user
+        if (currentUser.hasPermission(userId, SelfPermission.REPLACE_SELF_USER, Permission.REPLACE_USER)) {
+            userFacade.replaceUser(userId, userDTO);
+        } else
+            throw new GroupUnauthorizedException(Errors.UNAUTHORIZED_GROUP.getErrorMessage(currentUser.getGroup().getName()));
     }
 
     @Override
     @PostMapping("/{userId}/group")
     public void setGroup(@PathVariable UUID userId, @RequestBody GroupStrategy groupStrategy) {
-        userFacade.setUserGroup(userId, groupStrategy);
+        User currentUser = userFacade.getAuthenticatedUser();
+
+        // Checks if currentUser has permission to assign group to that user
+        if (currentUser.hasPermission(Permission.ASSIGN_GROUP)) {
+            userFacade.setUserGroup(userId, groupStrategy);
+        } else
+            throw new GroupUnauthorizedException(Errors.UNAUTHORIZED_GROUP.getErrorMessage(currentUser.getGroup().getName()));
     }
 }
